Documenting additional schema properties
I've added a brief note on the additional schema properties we use, and a link to the OS4D handbook where this is documented in more detail. 

If preferred, I can duplicate content from OS4D here - but was trying to stick to principle of keeping things in one place only.